requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
opencv-python==4.2.0.32
openpyxl==3.0.3
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
